reorganized methods, updated current dishes, need to add more documentation. test that everything works
