+++ favorites.py
@@ -1,11 +1,14 @@
 import streamlit as st
-import methods
 from datetime import date
-import visualization_methods as vm
+
+import methods.dishes_log_methods as dl
+import methods.favorites_methods as f
+import methods.visualization_methods as v
+import methods.users_methods as u
 
 user_id = st.session_state.get("user_id")
 
-if not methods.check_id(user_id):
+if not u.check_id(user_id):
     col1, col2 = st.columns((0.3, 1.5))
     col1.image(image='crumb-the-goose.png')
     col2.header("Peckish")
@@ -18,7 +21,7 @@ if not methods.check_id(user_id):
 
 st.header("Favorites")
 
-favs = methods.display_favorites(user_id)
+favs = f.display_favorites(user_id)
 
 if not favs.empty:
     st.write("---")
@@ -43,22 +46,22 @@ if not favs.empty:
                             key=f"notif {index} {row['dish_id']}"
                             )
         if toggle:
-            methods.toggle_notif(user_id, row['dish_id'])
+            f.toggle_notif(user_id, row['dish_id'])
             st.rerun()
         
-        col2.write(methods.get_dish_name(row["dish_id"]))
+        col2.write(dl.get_dish_name(row["dish_id"]))
         col3.write(row["date_added"])
 
         delete = col4.button("**-**", key=f"F{index}")
         if delete:
-            methods.remove_favorite(user_id, row["dish_id"])
+            f.remove_favorite(user_id, row["dish_id"])
             st.toast("Favorite Deleted")
             st.rerun()
 
     st.write("---")
 
     st.header("Favorites This Week")
-    faves = methods.get_faves_for_week(user_id, date.today())
+    faves = f.get_faves_for_week(user_id, date.today())
 
     if faves:
         st.write("---")
@@ -73,11 +76,11 @@ if not favs.empty:
         for fave in faves:
             col1, col2, col3, col4 = st.columns((1.75,0.75,0.75,0.75))
             col1.write(fave)
-            for f in faves[fave]:
+            for fav in faves[fave]:
                     
-                col2.write(f["meal"])
-                col3.write(f["location"])
-                col4.write("/".join(f["date"].split("-")[1:]))
+                col2.write(fav["meal"])
+                col3.write(fav["location"])
+                col4.write("/".join(fav["date"].split("-")[1:]))
             st.write("---")
     else:
         st.warning("Please turn on notifications for atleast one dish to see when it will be served.")
@@ -91,7 +94,7 @@ if not favs.empty:
     with col1.container():
         c1, c2 = st.columns((0.25, 3))
         i = 1
-        favorites = methods.weeklyTop5favs()
+        favorites = f.weeklyTop5favs()
         for fav in favorites:
             if fav[1] != 0:
                 c1.write(f"**{i}.**")
@@ -99,7 +102,7 @@ if not favs.empty:
                 i += 1
     col2.subheader("Top Dining Halls")
     with col2.container():
-                    st.bar_chart(vm.hall_popularity_last_7_days(), 
+                    st.bar_chart(v.hall_popularity_last_7_days(), 
                      horizontal=True, 
                      height=240,
                      width=250,
