+++ settings.py
@@ -1,9 +1,12 @@
 import streamlit as st
-import methods
+
+import methods.dishes_log_methods as dl
+import methods.friends_requests_methods as fr
+import methods.users_methods as u
 
 user_id = st.session_state.get("user_id")
 
-if not methods.check_id(user_id):
+if not u.check_id(user_id):
     col1, col2 = st.columns((0.3, 1.5))
     col1.image(image='crumb-the-goose.png')
     col2.header("Peckish")
@@ -18,7 +21,7 @@ st.header("Settings")
 st.divider()
 
 # === ALLERGENS AND PREFERENCES === #
-data = methods.get_user_allergens_preferences(user_id)
+data = u.get_user_allergens_preferences(user_id)
 if data:
     userA = data[0]
     userP = data[1]
@@ -58,7 +61,7 @@ try:
             
         st.write("")
         if st.button("Update"):
-                methods.update_user_allergens_preferences(user_id, selected_a, selected_p)
+                u.update_user_allergens_preferences(user_id, selected_a, selected_p)
 
         st.write("* Selections will filter menus shown in the Menus tab.")
 
@@ -71,33 +74,33 @@ try:
     col1, col2, col3 = st.columns((1,0.25,1))
     
     def toggle_optin(user_id):
-        methods.toggle_optin(user_id)
+        fr.toggle_optin(user_id)
 
     with col1:
             st.subheader("Activate Social")
-            social = st.checkbox("Opt-In", value = methods.get_optin(user_id), on_change=toggle_optin, args=(user_id,))
+            social = st.checkbox("Opt-In", value = fr.get_optin(user_id), on_change=toggle_optin, args=(user_id,))
             st.write("By opting into Social, you become searchable and can be added by others as a friend.")
             st.write("Users are searchable by username. You can change your username at any time.")
 
     with col3:
         st.subheader("Update Username")
         with st.form("Update Username"):
-            newname = st.text_input("New Username:", label_visibility="hidden", autocomplete=methods.get_username(user_id))
+            newname = st.text_input("New Username:", label_visibility="hidden", autocomplete=u.get_username(user_id))
             submitted = st.form_submit_button("Update")
             if submitted:
                 if newname == None or newname == "" or str.isspace(newname):
                     st.warning("Blank username not allowed.")
                 else:
-                    methods.set_username(user_id, newname)
-        st.write(f"**Username**: {methods.get_username(user_id)}")
+                    u.set_username(user_id, newname)
+        st.write(f"**Username**: {u.get_username(user_id)}")
     
     st.divider()
 
     # === EXTRA === #
     st.subheader("Usage Data")
-    st.write(f"Date joined: {methods.get_user_join_date(user_id)}")
-    st.write(f"Total dishes logged: {methods.get_total_dishes_logged(user_id)}")
-    st.write(f"Total friends: {len(methods.list_friends(user_id))}")
+    st.write(f"Date joined: {u.get_user_join_date(user_id)}")
+    st.write(f"Total dishes logged: {dl.get_total_dishes_logged(user_id)}")
+    st.write(f"Total friends: {len(fr.list_friends(user_id))}")
     
     st.divider()
